+++ src/BlockHorizons/BlockSniper/brush/registration/TypeRegistration.php
@@ -20,6 +20,7 @@ use BlockHorizons\BlockSniper\brush\types\MeltType;
 use BlockHorizons\BlockSniper\brush\types\OverlayType;
 use BlockHorizons\BlockSniper\brush\types\RegenerateType;
 use BlockHorizons\BlockSniper\brush\types\ReplaceAllType;
+use BlockHorizons\BlockSniper\brush\types\ReplaceTargetType;
 use BlockHorizons\BlockSniper\brush\types\ReplaceType;
 use BlockHorizons\BlockSniper\brush\types\SmoothType;
 use BlockHorizons\BlockSniper\brush\types\SnowConeType;
@@ -63,6 +64,7 @@ class TypeRegistration{
 		self::registerType(WarmType::class, WarmType::ID);
 		self::registerType(HeatType::class, HeatType::ID);
 		self::registerType(SmoothType::class, SmoothType::ID);
+		self::registerType(ReplaceTargetType::class, ReplaceTargetType::ID);
 	}
 
 	/**