+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Added a message if brush/obsolete blocks could not be resolved instead of defaulting to 0 silently.
 - The duration a modification took is now sent upon the finishing of the modification. Note that this is not entirely
   accurate, in particular for smaller modifications.
+- Added a new Replace Target type which replaces all blocks with the same type as the target block with the blocks set
+  as brush blocks.
   
 ### Changed
 - The Brush UI no longer contains all available brush properties.